+++ typescript/action_kit_api/CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
+## 2.1.0
+
+ - added support for target selection templates
+
 ## 2.0.1
 
-- Improved error-handling of extensions. Extensions are now able end with `failed` or `errored`. Extensions can also return an error within any response without
+ - Improved error-handling of extensions. Extensions are now able end with `failed` or `errored`. Extensions can also return an error within any response without
   the need to use a http status code != `200`. You can find more detailed descriptions about extension error handling in the docs.
 
 ## 2.0.0